x86/mm: x86/sgx: Propagate PF_SGX from error code to fixup_exception trapnr
Page faults with PF_SGX set are special, in every sense of the word.

In a bare metal environment, the PF_SGX bit is set in the error code
if and only if the fault was generated by the Enclave Page Cache Map
(EPCM).  In a virtualized environment, a VMM may also set the PF_SGX
bit when injecting an EPC related #PF.

Like normal page faults, PF_SGX can be set for either an invalid,
e.g. not-present, entry or for a fault due to a permission violation
on a valid entry, e.g. not-writable.  Because the SGX architecture
essentially gives hardware, and by extension VMMs, the freedom to
invalidate the EPC at any time and without warning, software must
be prepared to handle a page fault on any EPC access.  And because
such faults are by definition not due to the kernel's page tables,
and because the EPCM (or VMM) is completely opaque, the fault cannot
be resolved by the memory manager alone (unless SGX support is melded
into the memory manager).

This means that a properly functioning and bug-free SGX driver can
encounter page faults on kernel addresses on any given EPC access,
i.e. on any ENCLS instruction.  The Linux SGX driver will handle
such behavior by ensuring all ENCLS instructions are executed with
an associated ex_handler_fault().  But, the trap number alone does
not allow the driver to differentiate between a benign fault, e.g.
due to loss of EPC, and a fault due to a driver bug.

So, propagate the PF_SGX bit from the error code to the trapnr that
is passed to ex_handler_fault() so that the SGX driver can filter
out "benign" faults without having to suppress all WARNings on ENCLS
page faults.  Because trapnr is effectively capped at 31 on x86,
i.e. uses bits[4:0], and PF_SGX uses bit 15, we can smush in PF_SGX
without colliding with the actual trapnr.  And since PF_SGX is only
set on EPC accesses, only an entity deliberately interacting with the
EPC, e.g. SGX driver, will trigger faults with PF_SGX set in trapnr.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>